updated portfolio and nav
